# drm_appointment/models/end_appointment_data.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from drm_appointment.models.base_model_ import Model
from drm_appointment import util


class EndAppointmentData(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, appointment_id=None):  # noqa: E501
        """EndAppointmentData - a model defined in OpenAPI

        :param appointment_id: The appointment_id of this EndAppointmentData.  # noqa: E501
        :type appointment_id: str
        """
        self.openapi_types = {"appointment_id": str}

        self.attribute_map = {"appointment_id": "appointment_id"}

        self._appointment_id = appointment_id

    @classmethod
    def from_dict(cls, dikt) -> "EndAppointmentData":
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The EndAppointmentData of this EndAppointmentData.  # noqa: E501
        :rtype: EndAppointmentData
        """
        return util.deserialize_model(dikt, cls)

    @property
    def appointment_id(self):
        """Gets the appointment_id of this EndAppointmentData.


        :return: The appointment_id of this EndAppointmentData.
        :rtype: str
        """
        return self._appointment_id

    @appointment_id.setter
    def appointment_id(self, appointment_id):
        """Sets the appointment_id of this EndAppointmentData.


        :param appointment_id: The appointment_id of this EndAppointmentData.
        :type appointment_id: str
        """

        self._appointment_id = appointment_id