# drm_appointment/models/patient_post_data.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from drm_appointment.models.base_model_ import Model
from drm_appointment import util


class PatientPostData(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, name=None):  # noqa: E501
        """PatientPostData - a model defined in OpenAPI

        :param name: The name of this PatientPostData.  # noqa: E501
        :type name: str
        """
        self.openapi_types = {"name": str}

        self.attribute_map = {"name": "name"}

        self._name = name

    @classmethod
    def from_dict(cls, dikt) -> "PatientPostData":
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The PatientPostData of this PatientPostData.  # noqa: E501
        :rtype: PatientPostData
        """
        return util.deserialize_model(dikt, cls)

    @property
    def name(self):
        """Gets the name of this PatientPostData.


        :return: The name of this PatientPostData.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name):
        """Sets the name of this PatientPostData.


        :param name: The name of this PatientPostData.
        :type name: str
        """
        if name is None:
            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501

        self._name = name