# drm_appointment/models/patient_referal_id.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from drm_appointment.models.base_model_ import Model
from drm_appointment import util


class PatientReferalId(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, id=None):  # noqa: E501
        """PatientReferalId - a model defined in OpenAPI

        :param id: The id of this PatientReferalId.  # noqa: E501
        :type id: str
        """
        self.openapi_types = {"id": str}

        self.attribute_map = {"id": "id"}

        self._id = id

    @classmethod
    def from_dict(cls, dikt) -> "PatientReferalId":
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The PatientReferalId of this PatientReferalId.  # noqa: E501
        :rtype: PatientReferalId
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self):
        """Gets the id of this PatientReferalId.


        :return: The id of this PatientReferalId.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id):
        """Sets the id of this PatientReferalId.


        :param id: The id of this PatientReferalId.
        :type id: str
        """

        self._id = id