# drm_appointment/models/physician_referal_id.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from drm_appointment.models.base_model_ import Model
from drm_appointment import util


class PhysicianReferalId(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, id=None):  # noqa: E501
        """PhysicianReferalId - a model defined in OpenAPI

        :param id: The id of this PhysicianReferalId.  # noqa: E501
        :type id: str
        """
        self.openapi_types = {"id": str}

        self.attribute_map = {"id": "id"}

        self._id = id

    @classmethod
    def from_dict(cls, dikt) -> "PhysicianReferalId":
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The PhysicianReferalId of this PhysicianReferalId.  # noqa: E501
        :rtype: PhysicianReferalId
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self):
        """Gets the id of this PhysicianReferalId.


        :return: The id of this PhysicianReferalId.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id):
        """Sets the id of this PhysicianReferalId.


        :param id: The id of this PhysicianReferalId.
        :type id: str
        """

        self._id = id