# drm_appointment/models/start_appointment_data.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from drm_appointment.models.base_model_ import Model
from drm_appointment import util


class StartAppointmentData(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, physician_id=None, patient_id=None):  # noqa: E501
        """StartAppointmentData - a model defined in OpenAPI

        :param physician_id: The physician_id of this StartAppointmentData.  # noqa: E501
        :type physician_id: str
        :param patient_id: The patient_id of this StartAppointmentData.  # noqa: E501
        :type patient_id: str
        """
        self.openapi_types = {"physician_id": str, "patient_id": str}

        self.attribute_map = {"physician_id": "physician_id", "patient_id": "patient_id"}

        self._physician_id = physician_id
        self._patient_id = patient_id

    @classmethod
    def from_dict(cls, dikt) -> "StartAppointmentData":
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The StartAppointmentData of this StartAppointmentData.  # noqa: E501
        :rtype: StartAppointmentData
        """
        return util.deserialize_model(dikt, cls)

    @property
    def physician_id(self):
        """Gets the physician_id of this StartAppointmentData.


        :return: The physician_id of this StartAppointmentData.
        :rtype: str
        """
        return self._physician_id

    @physician_id.setter
    def physician_id(self, physician_id):
        """Sets the physician_id of this StartAppointmentData.


        :param physician_id: The physician_id of this StartAppointmentData.
        :type physician_id: str
        """

        self._physician_id = physician_id

    @property
    def patient_id(self):
        """Gets the patient_id of this StartAppointmentData.


        :return: The patient_id of this StartAppointmentData.
        :rtype: str
        """
        return self._patient_id

    @patient_id.setter
    def patient_id(self, patient_id):
        """Sets the patient_id of this StartAppointmentData.


        :param patient_id: The patient_id of this StartAppointmentData.
        :type patient_id: str
        """

        self._patient_id = patient_id